make necessary changes to allow Wait to handle non-target expectations.
This is one huge-honkin' commit.

The original task was small -- allow Wait to be able to use conditions that
didn't use a target, like url_changes. But while figuring out how to handle
that, i realized that BrowseTheWeb was doing too many things, and limiting
what Wait would be able to do because of how prescriptive its to_wait_for
method was.

Besides violating the S of SOLID, the extra trappings were annoying to deal
with. So i ripped them all out. And the to_find parts, too, which got rid of a
circular import that has been bothering me for a while.

In the end, we've got updated tests, updated docs, and yowza a big paradigm
shift in abilities. Phew!