remove unnecessary complication for Prop.
While attempting to switch over to a singleton Narrator, i thought i had
run into an issue that required us to define the Prop class inside of the
test functions that used it. It turned out i had a bug in the Narrator
class itself that was causing the weird behavior, so these changes were
no longer necessary once i fixed that bug!